syz-cluster/controller: move services to pkg/service
This will facilitate the reuse of the code.
Split off SessionService from SeriesService.
